requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pymysql==1.1.0
flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
lxml==4.9.3
schedule==1.2.0
python-dotenv==1.0.0
selenium==4.15.0
pandas==2.1.3
numpy==1.24.3